[CRITICAL] Fixed import manager issue.
